pyproject.toml
[tool.poetry]
name = "samaraenergo"
version = "0.2.25"
description = "Клиент электросбытовой компании СамараЭнерго"
authors = ["Sergey V. DUDANOV <sergey.dudanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
aiohttp = ">=3.11"
pydantic = ">=2.10.4"
pymupdf = "^1.25.1"
pillow = "^11.0.0"
tzdata = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
